+++ src/base/types_def.h
@@ -215,14 +215,6 @@ typedef enum _app_type_t {
 #define WITH_VGCANVAS 1
 #endif /*defined(WITH_NANOVG_SOFT) || defined(WITH_NANOVG_GPU)*/
 
-#ifndef TK_DEFAULT_FONT
-#ifdef WITH_MINI_FONT
-#define TK_DEFAULT_FONT "default_mini"
-#else
-#define TK_DEFAULT_FONT "default"
-#endif /*WITH_MINI_FONT*/
-#endif /*TK_DEFAULT_FONT*/
-
 #ifndef TK_DEFAULT_FONT_SIZE
 #define TK_DEFAULT_FONT_SIZE 18
 #endif /*TK_DEFAULT_FONT_SIZE*/